+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI,INTERNAL: Provide WithProfileSelection via koin
 - SDK,INTERNAL: Separate Uia logic from UiaRouter
 - SDK,INTERNAL: Separate SendAttachmentRouter from TimelineViewModel
+- UI: made several compose functions for notifications public instead of internal
 
 ### Deprecated